Adds MLD model vs obs python script
This adds the mld vs obs comparison notebook into the new framework.  In
the process some modularization was introduced.  Interpolation of data
was placed into shared and computing of climatologies is simplified.

renames file for consistency

Changed module name in run_analysis but did not change file name itself

Fixes a number of bugs

Fixes issues in the implementation of the new additions

Fixes a labeling issue and cleans up unneeded code

The MLD plots were still labeled SST, which has been changed.

I also cleaned out some unneeded code

Reverts config.analysis to defaults where appropriate
